Extract print_invocations
